+++ packages/python/qsu/os/getFreeRamSize.py
@@ -2,10 +2,10 @@ from ..format.fileSizeFormat import fileSizeFormat
 from ._system import memorySize
 
 
-def getRamSize() -> str:
+def getFreeRamSize() -> str:
 	memory = memorySize()
 
 	if memory is None:
 		raise RuntimeError('Failed to read the size of the physical memory')
 
-	return fileSizeFormat(memory[0], 0, True)
+	return fileSizeFormat(memory[1], 0, True)