+++ packages/python/qsu/os/getUsedRamSize.py
@@ -2,10 +2,12 @@ from ..format.fileSizeFormat import fileSizeFormat
 from ._system import memorySize
 
 
-def getRamSize() -> str:
+def getUsedRamSize() -> str:
 	memory = memorySize()
 
 	if memory is None:
 		raise RuntimeError('Failed to read the size of the physical memory')
 
-	return fileSizeFormat(memory[0], 0, True)
+	total, free = memory
+
+	return fileSizeFormat(max(total - free, 0), 0, True)